requirements.txt
gradio
qdrant_client
langchain
langchain-community
langchain_experimental
llama-parse
fastembed
#chromadb
python-dotenv
#langchain-groq
chainlit
unstructured[md]